task 5

+++ bit_manipulation/5-flip_bits.c
@@ -4,7 +4,7 @@
  *
  *
  */
-unsigned int countSetBits(int n)
+unsigned int countSetBits(unsigned int long n)
 {
     unsigned int count = 0;
     while (n > 0) {
